requirements.txt
albumentations==1.1.0
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.62
Pillow==8.4.0
scikit-image==0.19.1
scikit-learn==1.0.2
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2

